Fix unsigned integer arithmetic bug
git-svn-id: http://svn.code.sf.net/p/netpbm/code/trunk@3063 9d0c8265-081b-0410-96cb-a4ca84ce46f8

+++ editor/ppmbrighten.c
@@ -102,7 +102,7 @@ parseCommandLine(int argc, char ** argv,
 static __inline__ unsigned int
 mod(int const dividend, unsigned int const divisor) {
 
-    int remainder = dividend % divisor;
+    int remainder = dividend % (int)divisor;
 
     if (remainder < 0)
         return divisor + remainder;
